# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AlertAckRequest:
    properties:
      comment:
        maxLength: 500
        type: string
    type: object
...
  dto.AlertRuleListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AlertRuleResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  dto.AlertSilenceListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.AlertSilenceResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:10010
info:
  contact:
    name: TicketDesk
    url: https://github.com/kerbos/ticketdesk
  description: |-
    项目化工单与告警联动系统 API. 一切问题都是工单, 一切告警都必须被跟进.
    **统一响应格式**: `{code, message, data}`. 错误响应 `{code, message, details?}` (code 为字符串如 "BAD_REQUEST" / "UNAUTHORIZED").
    **速率限制**: 全站默认 300 req/min/IP, /auth/* 端点 20 req/min/IP, /webhook/* 100 req/min/IP. 超出返回 **429 Too Many Requests**, 此状态码在所有端点皆可能出现, 单独 endpoint 文档不重复声明.
  license:
    name: MIT
  termsOfService: https://github.com/kerbos/ticketdesk
  title: TicketDesk API
  version: "1.0"
paths:
...
  /api/v1/alert-rules:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertRuleListResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警规则列表
      tags:
      - Alert
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建告警规则请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAlertRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertRuleResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建告警规则
      tags:
      - Alert
  /api/v1/alert-rules/{id}:
    delete:
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除告警规则
      tags:
      - Alert
    get:
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertRuleResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 规则不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警规则详情
      tags:
      - Alert
    put:
      consumes:
      - application/json
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新告警规则请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAlertRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertRuleResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新告警规则
      tags:
      - Alert
  /api/v1/alert-silences:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertSilenceListResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警静默列表
      tags:
      - Alert
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建静默请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAlertSilenceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertSilenceResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建告警静默
      tags:
      - Alert
  /api/v1/alert-silences/{id}:
    delete:
      parameters:
      - description: 静默 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除告警静默
      tags:
      - Alert
    get:
      parameters:
      - description: 静默 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertSilenceResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 静默不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警静默详情
      tags:
      - Alert
    put:
      consumes:
      - application/json
      parameters:
      - description: 静默 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新静默请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAlertSilenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertSilenceResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新告警静默
      tags:
      - Alert
  /api/v1/alert-silences/{id}/cancel:
    post:
      parameters:
      - description: 静默 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 取消成功
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 取消告警静默
      tags:
      - Alert
  /api/v1/alerts:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 告警状态 (firing/resolved)
        in: query
        name: status
        type: string
      - description: 严重级别 (critical/warning/info)
        in: query
        name: severity
        type: string
      - description: 来源
        in: query
        name: source
        type: string
      - description: 告警名称
        in: query
        name: alert_name
        type: string
      - description: 关联工单 ID
        in: query
        name: issue_id
        type: integer
      - description: '标签筛选，格式: key==value,key!=value'
        in: query
        name: label_filters
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertListResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警列表
      tags:
      - Alert
  /api/v1/alerts/{id}:
    get:
      parameters:
      - description: 告警 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AlertResponse'
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 告警不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警详情
      tags:
      - Alert
  /api/v1/alerts/{id}/ack:
    post:
      consumes:
      - application/json
      parameters:
      - description: 告警 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 确认告警请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AlertAckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 确认成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 确认告警
      tags:
      - Alert
  /api/v1/alerts/{id}/resolve:
    post:
      consumes:
      - application/json
      parameters:
      - description: 告警 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 解决告警请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AlertResolveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 解决成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 解决告警
      tags:
      - Alert
...
  /api/v1/alerts/group:
    get:
      parameters:
      - description: 分组字段（标签 key）
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.AlertGroupResponse'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 按标签分组统计告警
      tags:
      - Alert
  /api/v1/alerts/label-keys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警标签键列表
      tags:
      - Alert
  /api/v1/alerts/nightingale:
    post:
      consumes:
      - application/json
      description: 接收来自夜莺监控系统的告警推送，支持单个或数组格式，无需认证
      parameters:
      - description: 夜莺告警事件（数组或单个对象）
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/dto.N9eAlertEvent'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 处理成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 接收夜莺告警 Webhook
      tags:
      - Alert
  /api/v1/alerts/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功（data 为 AlertStatsResponse）
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取告警统计数据
      tags:
      - Alert
  /api/v1/alerts/webhook:
    post:
      consumes:
      - application/json
      description: 接收来自 Alertmanager 的告警推送，无需认证
      parameters:
      - description: 告警 Webhook 请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AlertWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 处理成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: 接收 Prometheus 告警 Webhook
      tags:
      - Alert
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer 令牌 (`Bearer <jwt>`) 或 Personal Access Token (`Bearer td_pat_xxx`)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"